Replaced FirebaseJobDispatcher with WorkManager.
