+++ apps/teacher/src/main/java/com/instructure/teacher/activities/SignInActivity.kt
@@ -23,14 +23,13 @@ import android.webkit.CookieManager
 import com.instructure.canvasapi2.models.AccountDomain
 import com.instructure.canvasapi2.utils.ApiPrefs.isMasquerading
 import com.instructure.loginapi.login.activities.BaseLoginSignInActivity
-import com.instructure.pandautils.services.PushNotificationRegistrationService.Companion.scheduleJob
-import com.instructure.teacher.BuildConfig
+import com.instructure.pandautils.services.PushNotificationRegistrationWorker
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class SignInActivity : BaseLoginSignInActivity() {
     override fun launchApplicationMainActivityIntent(): Intent {
-        scheduleJob(this, isMasquerading)
+        PushNotificationRegistrationWorker.scheduleJob(this, isMasquerading)
         CookieManager.getInstance().flush()
         return SplashActivity.createIntent(this, null)
     }
